[Fix] Put all admin components under one folder
- Put all admin table/visual folders in components under one folder named "admin" in components.
- Added comments for statCards files.
