+++ services/frontend-v3/src/components/admin/competencyTable/CompetencyTableView.jsx
