+++ services/frontend-v3/src/components/admin/diplomaTable/DiplomaTable.jsx
@@ -1,19 +1,19 @@
 import React, { useState, useEffect } from "react";
-import CategoryTableView from "./CategoryTableView";
+import DiplomaTableView from "./DiplomaTableView";
 import { Skeleton } from "antd";
 import axios from "axios";
 import _ from "lodash";
 import { injectIntl } from "react-intl";
-import config from "../../config";
+import config from "../../../config";
 
 const backendAddress = config.backendAddress;
 
 /**
- *  CategoryTable(props)
- *  Controller for the CategoryTableView.
+ *  DiplomaTable(props)
+ *  Controller for the DiplomaTableView.
  *  It gathers the required data for rendering the component.
  */
-function CategoryTable(props) {
+function DiplomaTable(props) {
   const [data, setData] = useState([]);
   const [loading, setLoading] = useState(true);
   const [reset, setReset] = useState(false);
@@ -27,89 +27,31 @@ function CategoryTable(props) {
   /* useEffect will run if statement, when the component is mounted */
   /* useEffect will run else statement, if an addition, update/edit or deletion occurs in the table */
   useEffect(() => {
-    let categories = [];
+    let diplomas = [];
     if (loading) {
       const setState = async () => {
-        categories = await getCategories();
-        setData(categories);
+        diplomas = await getDiplomas();
+        setData(diplomas);
         setLoading(false);
       };
       setState();
     } else {
       const updateState = async () => {
-        categories = await getCategories();
-        setData(categories);
+        diplomas = await getDiplomas();
+        setData(diplomas);
         setReset(false);
-        setSelectedRowKeys([]);
       };
       updateState();
     }
   }, [loading, reset]);
 
-  /* get category information */
-  const getCategories = async () => {
+  /* get diploma information */
+  const getDiplomas = async () => {
     try {
       let results = await axios.get(
-        backendAddress + "api/admin/options/categories/skill"
+        backendAddress + "api/admin/options/" + type
       );
       return results.data;
-    } catch (error) {
-      console.log(error);
-      return 0;
-    }
-  };
-
-  /* handles the deletion of a category */
-  const handleSubmitDelete = async () => {
-    try {
-      const url = backendAddress + "api/admin/delete/" + type;
-
-      let result;
-
-      await axios.post(url, { ids: selectedRowKeys }).then(function (response) {
-        result = response.data.deletePerformed;
-      });
-
-      if (result === false) {
-        return true;
-      } else {
-        setReset(true);
-        return false;
-      }
-    } catch (error) {
-      console.log(error);
-      return 0;
-    }
-  };
-
-  /* handles addition of a category */
-  const handleSubmitAdd = async (values) => {
-    try {
-      const url = backendAddress + "api/admin/options/" + type;
-
-      await axios.post(url, {
-        descriptionEn: values.addCategoryEn,
-        descriptionFr: values.addCategoryFr,
-      });
-
-      setReset(true);
-    } catch (error) {
-      console.log(error);
-      return 0;
-    }
-  };
-
-  /* handles the update/edit of a category */
-  const handleSubmitEdit = async (values, id) => {
-    try {
-      const url = backendAddress + "api/admin/options/" + type + "/" + id;
-
-      await axios.put(url, {
-        descriptionEn: values.editCategoryEn,
-        descriptionFr: values.editCategoryFr,
-      });
-
-      setReset(true);
     } catch (error) {
       console.log(error);
       return 0;
@@ -145,6 +87,55 @@ function CategoryTable(props) {
     setSearchText("");
   };
 
+  /* handles addition of a diploma */
+  const handleSubmitAdd = async (values) => {
+    try {
+      const url = backendAddress + "api/admin/options/" + type;
+
+      await axios.post(url, {
+        descriptionEn: values.addDiplomaEn,
+        descriptionFr: values.addDiplomaFr,
+      });
+
+      setReset(true);
+    } catch (error) {
+      console.log(error);
+      return 0;
+    }
+  };
+
+  /* handles the update/edit of a diploma */
+  const handleSubmitEdit = async (values, id) => {
+    try {
+      const url = backendAddress + "api/admin/options/" + type + "/" + id;
+
+      await axios.put(url, {
+        descriptionEn: values.editDiplomaEn,
+        descriptionFr: values.editDiplomaFr,
+      });
+
+      setReset(true);
+    } catch (error) {
+      console.log(error);
+      return 0;
+    }
+  };
+
+  /* handles the deletion of a diploma */
+  const handleSubmitDelete = async () => {
+    try {
+      const url = backendAddress + "api/admin/delete/" + type;
+
+      await axios.post(url, { ids: selectedRowKeys });
+
+      setSelectedRowKeys([]);
+      setReset(true);
+    } catch (error) {
+      console.log(error);
+      return 0;
+    }
+  };
+
   /* handles row selection in the table */
   // Consult: function taken from Ant Design table components (updated to functional)
   const rowSelection = {
@@ -156,25 +147,25 @@ function CategoryTable(props) {
   /* helper function to rowSelection */
   // Consult: function taken from Ant Design table components (updated to functional)
   const onSelectChange = (selectedRowKeys) => {
-    // Can access the keys of each category selected in the table
+    // Can access the keys of each diploma selected in the table
     setSelectedRowKeys(selectedRowKeys);
   };
 
   /* configures data from backend into viewable data for the table */
-  const getCategoryInformation = () => {
+  const convertToViewableInformation = () => {
     // Allows for sorting of data between French/English in terms of description:
     const description =
       props.intl.formatMessage({ id: "language.code" }) === "en"
         ? "descriptionEn"
         : "descriptionFr";
 
-    let allCategories = _.sortBy(data, description);
+    let allDiplomas = _.sortBy(data, description);
 
-    for (let i = 0; i < allCategories.length; i++) {
-      allCategories[i].key = allCategories[i].id;
+    for (let i = 0; i < allDiplomas.length; i++) {
+      allDiplomas[i].key = allDiplomas[i].id;
     }
 
-    return allCategories;
+    return allDiplomas;
   };
 
   document.title = getDisplayType(true) + " - Admin | I-Talent";
@@ -184,7 +175,7 @@ function CategoryTable(props) {
   }
 
   return (
-    <CategoryTableView
+    <DiplomaTableView
       handleSearch={handleSearch}
       handleReset={handleReset}
       handleSubmitAdd={handleSubmitAdd}
@@ -195,9 +186,9 @@ function CategoryTable(props) {
       searchText={searchText}
       size={size}
       rowSelection={rowSelection}
-      data={getCategoryInformation()}
+      data={convertToViewableInformation()}
     />
   );
 }
 
-export default injectIntl(CategoryTable);
+export default injectIntl(DiplomaTable);
