+++ services/frontend-v3/src/components/admin/diplomaTable/DiplomaTableView.jsx
