+++ services/frontend-v3/src/components/admin/skillTable/SkillTable.jsx
@@ -1,20 +1,21 @@
 import React, { useState, useEffect } from "react";
-import SchoolTableView from "./SchoolTableView";
+import SkillTableView from "./SkillTableView";
 import { Skeleton } from "antd";
 import axios from "axios";
 import _ from "lodash";
 import { injectIntl } from "react-intl";
-import config from "../../config";
+import config from "../../../config";
 
 const backendAddress = config.backendAddress;
 
 /**
- *  SchoolTable(props)
- *  Controller for the SchoolTableView.
+ *  SkillTable(props)
+ *  Controller for the SkillTableView.
  *  It gathers the required data for rendering the component.
  */
-function SchoolTable(props) {
+function SkillTable(props) {
   const [data, setData] = useState([]);
+  const [categories, setCategories] = useState([]);
   const [loading, setLoading] = useState(true);
   const [reset, setReset] = useState(false);
   const [searchText, setSearchText] = useState("");
@@ -27,31 +28,46 @@ function SchoolTable(props) {
   /* useEffect will run if statement, when the component is mounted */
   /* useEffect will run else statement, if an addition, update/edit or deletion occurs in the table */
   useEffect(() => {
-    let schools = [];
+    let skills = [];
+    let categories = [];
     if (loading) {
       const setState = async () => {
-        schools = await getSchools();
-        setData(schools);
-        console.log("Before: ", schools);
+        skills = await getSkill();
+        categories = await getCategories();
+        setData(skills);
+        setCategories(categories);
         setLoading(false);
       };
       setState();
     } else {
       const updateState = async () => {
-        schools = await getSchools();
-        setData(schools);
-        console.log("After: ", schools);
+        skills = await getSkill();
+        setData(skills);
         setReset(false);
       };
       updateState();
     }
   }, [loading, reset]);
 
-  /* get school information */
-  const getSchools = async () => {
+  /* get skill information */
+  const getSkill = async () => {
     try {
       let results = await axios.get(
         backendAddress + "api/admin/options/" + type
+      );
+
+      return results.data;
+    } catch (error) {
+      console.log(error);
+      return 0;
+    }
+  };
+
+  /* get category information */
+  const getCategories = async () => {
+    try {
+      let results = await axios.get(
+        backendAddress + "api/admin/options/categories/" + type
       );
       return results.data;
     } catch (error) {
@@ -89,15 +105,15 @@ function SchoolTable(props) {
     setSearchText("");
   };
 
-  /* handles addition of a school */
+  /* handles addition of a skill */
   const handleSubmitAdd = async (values) => {
     try {
       const url = backendAddress + "api/admin/options/" + type;
 
       await axios.post(url, {
-        country: values.addSchoolCountry.toUpperCase(),
-        description: values.addSchoolName,
-        state: values.addSchoolState.toUpperCase(),
+        descriptionEn: values.addSkillEn,
+        descriptionFr: values.addSkillFr,
+        categoryId: values.addSkillCategory,
       });
 
       setReset(true);
@@ -107,16 +123,34 @@ function SchoolTable(props) {
     }
   };
 
-  /* handles the update/edit of a school */
+  /* handles the update/edit of a skill */
   const handleSubmitEdit = async (values, id) => {
     try {
       const url = backendAddress + "api/admin/options/" + type + "/" + id;
 
-      await axios.put(url, {
-        country: values.editSchoolCountry.toUpperCase(),
-        description: values.editSchoolName,
-        state: values.editSchoolState.toUpperCase(),
-      });
+      if (typeof values.editSkillCategory === "string") {
+        const index = categories.findIndex(
+          (object) =>
+            object.descriptionEn === values.editSkillCategory ||
+            object.descriptionFr === values.editSkillCategory
+        );
+        await axios.put(url, {
+          descriptionEn: values.editSkillEn,
+          descriptionFr: values.editSkillFr,
+          categoryId: categories[index].id,
+          category: categories[index],
+        });
+      } else {
+        let categoryObject = categories.find(
+          (category) => category.id === values.editSkillCategory
+        );
+        await axios.put(url, {
+          descriptionEn: values.editSkillEn,
+          descriptionFr: values.editSkillFr,
+          categoryId: values.editSkillCategory,
+          category: categoryObject,
+        });
+      }
 
       setReset(true);
     } catch (error) {
@@ -125,7 +159,7 @@ function SchoolTable(props) {
     }
   };
 
-  /* handles the deletion of a school */
+  /* handles the deletion of a skill */
   const handleSubmitDelete = async () => {
     try {
       const url = backendAddress + "api/admin/delete/" + type;
@@ -151,19 +185,35 @@ function SchoolTable(props) {
   /* helper function to rowSelection */
   // Consult: function taken from Ant Design table components (updated to functional)
   const onSelectChange = (selectedRowKeys) => {
-    // Can access the keys of each school selected in the table
+    // Can access the keys of each skill selected in the table
     setSelectedRowKeys(selectedRowKeys);
   };
 
   /* configures data from backend into viewable data for the table */
-  const convertToViewableInformation = () => {
-    let allSchools = _.sortBy(data, "description");
+  const getSkillInformation = () => {
+    // Allows for sorting of data between French/English in terms of description and category:
+    const description =
+      props.intl.formatMessage({ id: "language.code" }) === "en"
+        ? "descriptionEn"
+        : "descriptionFr";
 
-    for (let i = 0; i < allSchools.length; i++) {
-      allSchools[i].key = allSchools[i].id;
+    const category =
+      props.intl.formatMessage({ id: "language.code" }) === "en"
+        ? "categoryNameEn"
+        : "categoryNameFr";
+
+    let allSkills = _.sortBy(data, description);
+
+    for (let i = 0; i < allSkills.length; i++) {
+      allSkills[i].key = allSkills[i].id;
     }
 
-    return allSchools;
+    allSkills.forEach((e) => {
+      e.categoryNameEn = e.category.descriptionEn;
+      e.categoryNameFr = e.category.descriptionFr;
+    });
+
+    return _.sortBy(allSkills, category);
   };
 
   document.title = getDisplayType(true) + " - Admin | I-Talent";
@@ -173,7 +223,7 @@ function SchoolTable(props) {
   }
 
   return (
-    <SchoolTableView
+    <SkillTableView
       handleSearch={handleSearch}
       handleReset={handleReset}
       handleSubmitAdd={handleSubmitAdd}
@@ -184,9 +234,10 @@ function SchoolTable(props) {
       searchText={searchText}
       size={size}
       rowSelection={rowSelection}
-      data={convertToViewableInformation()}
+      data={getSkillInformation()}
+      categories={categories}
     />
   );
 }
 
-export default injectIntl(SchoolTable);
+export default injectIntl(SkillTable);
