+++ services/frontend-v3/src/components/admin/skillTable/SkillTableView.jsx
