+++ services/frontend-v3/src/components/admin/statCards/StatCards.jsx
@@ -1,8 +1,16 @@
 import React from "react";
 import StatCardsView from "./StatCardsView";
 
+/**
+ *  StatCards(props)
+ *  Controller for the StatCardsView.
+ *  It setups the data (bridge) for rendering the component in the view.
+ */
 function StatCards(props) {
-  const dashboardCount = data => {
+  const data = props.data;
+  /* gets data for the stat cards on dashboard */
+  // Gets data for Total users, Inactive users, Hidden profiles, and Total Ex Feeders
+  const dashboardCount = (data) => {
     const total_users = data.user;
     const inactive_users = data.inactive;
     const hiddenProfiles = data.flagged;
@@ -10,15 +18,15 @@ function StatCards(props) {
     return { total_users, inactive_users, hiddenProfiles, exFeeders };
   };
 
-  const monthGrowthRate = data => {
+  /* gets monthly data for the stat cards on dashboard */
+  // New users in month, and Growth rate in month
+  const monthGrowthRate = (data) => {
     const growthRate = data.growthRateFromPreviousMonth;
 
     const current_month_additions = data.current_month_additions.count;
 
     return { growthRate, current_month_additions };
   };
-
-  const data = props.data;
 
   return (
     <StatCardsView
