+++ services/frontend-v3/src/pages/admin/Category.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import AdminLayout from "../../components/layouts/adminLayout/AdminLayout";
-import CategoryTable from "../../components/categoryTable/CategoryTable";
+import CategoryTable from "../../components/admin/categoryTable/CategoryTable";
 
 function AdminCategory(props) {
   return (
