+++ services/frontend-v3/src/pages/admin/Competency.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import AdminLayout from "../../components/layouts/adminLayout/AdminLayout";
-import CompetencyTable from "../../components/competencyTable/CompetencyTable";
+import CompetencyTable from "../../components/admin/competencyTable/CompetencyTable";
 
 function AdminCompetency(props) {
   return (
