+++ services/frontend-v3/src/pages/admin/School.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import AdminLayout from "../../components/layouts/adminLayout/AdminLayout";
-import SchoolTable from "../../components/schoolTable/SchoolTable";
+import SchoolTable from "../../components/admin/schoolTable/SchoolTable";
 
 function AdminSchool(props) {
   return (
